Remove 'I' from extension method classes and test classes.
