+++ src/OnionSeed.Data/AsyncRepositoryExtensions.cs
@@ -4,29 +4,29 @@ using OnionSeed.Data.Decorators;
 namespace OnionSeed.Data
 {
 	/// <summary>
-	/// Contains extension methods for <see cref="ICommandService{TEntity, TIdentity}"/>.
+	/// Contains extension methods for <see cref="IAsyncRepository{TEntity, TIdentity}"/>.
 	/// </summary>
-	public static class ICommandServiceExtensions
+	public static class AsyncRepositoryExtensions
 	{
 		/// <summary>
-		/// Wraps the given <see cref="ICommandService{TEntity, TIdentity}"/> in a <see cref="CommandServiceExceptionHandlerDecorator{TEntity, TIdentity, TException}"/>.
+		/// Wraps the given <see cref="IAsyncRepository{TEntity, TIdentity}"/> in a <see cref="AsyncRepositoryExceptionHandlerDecorator{TEntity, TIdentity, TException}"/>.
 		/// </summary>
 		/// <typeparam name="TEntity">The type of entities in the data store.</typeparam>
 		/// <typeparam name="TIdentity">The type of the unique identity value of the entities in the data store.</typeparam>
 		/// <typeparam name="TException">"The type of exception to be handled.</typeparam>
-		/// <param name="inner">The <see cref="ICommandService{TEntity, TIdentity}"/> to be wrapped.</param>
+		/// <param name="inner">The <see cref="IAsyncRepository{TEntity, TIdentity}"/> to be wrapped.</param>
 		/// <param name="handler">The handler that will be called when an exception is caught.
 		/// This delegate must return a flag indicating if the exception was handled.
 		/// If it wasn't, it will be re-thrown after processing.</param>
-		/// <returns>A new <see cref="CommandServiceExceptionHandlerDecorator{TEntity, TIdentity, TException}"/> wrapping the given <see cref="ICommandService{TEntity, TIdentity}"/>.</returns>
+		/// <returns>A new <see cref="AsyncRepositoryExceptionHandlerDecorator{TEntity, TIdentity, TException}"/> wrapping the given <see cref="IAsyncRepository{TEntity, TIdentity}"/>.</returns>
 		/// <exception cref="ArgumentNullException"><paramref name="inner"/> is <c>null</c>.
 		/// -or- <paramref name="handler"/> is <c>null</c>.</exception>
-		public static ICommandService<TEntity, TIdentity> Catch<TEntity, TIdentity, TException>(this ICommandService<TEntity, TIdentity> inner, Func<TException, bool> handler)
+		public static IAsyncRepository<TEntity, TIdentity> Catch<TEntity, TIdentity, TException>(this IAsyncRepository<TEntity, TIdentity> inner, Func<TException, bool> handler)
 			where TEntity : IEntity<TIdentity>
 			where TIdentity : IEquatable<TIdentity>, IComparable<TIdentity>
 			where TException : Exception
 		{
-			return new CommandServiceExceptionHandlerDecorator<TEntity, TIdentity, TException>(inner, handler);
+			return new AsyncRepositoryExceptionHandlerDecorator<TEntity, TIdentity, TException>(inner, handler);
 		}
 	}
 }
