+++ test/OnionSeed.Data.Tests/QueryServiceExtensionsTests.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Diagnostics.CodeAnalysis;
-using System.Threading.Tasks;
 using FluentAssertions;
 using Moq;
 using OnionSeed.Data.Decorators;
@@ -8,10 +6,9 @@ using Xunit;
 
 namespace OnionSeed.Data
 {
-	[SuppressMessage("AsyncUsage.CSharp.Naming", "UseAsyncSuffix:Use Async suffix", Justification = "Test methods don't need to end in 'Async'.")]
-	public class IAsyncCommandServiceExtensionsTests
+	public class QueryServiceExtensionsTests
 	{
-		private readonly Mock<IAsyncCommandService<FakeEntity<int>, int>> _mockInner = new Mock<IAsyncCommandService<FakeEntity<int>, int>>(MockBehavior.Strict);
+		private readonly Mock<IQueryService<FakeEntity<int>, int>> _mockInner = new Mock<IQueryService<FakeEntity<int>, int>>(MockBehavior.Strict);
 
 		[Theory]
 		[InlineData(false, false)]
@@ -32,16 +29,15 @@ namespace OnionSeed.Data
 		}
 
 		[Fact]
-		public async Task Catch_ShouldWrapInner()
+		public void Catch_ShouldWrapInner()
 		{
 			// Arrange
-			const int id = 5;
 			var expectedException = new InvalidOperationException();
 			InvalidOperationException actualException = null;
 
 			_mockInner
-				.Setup(i => i.TryRemoveAsync(id))
-				.ThrowsAsync(expectedException);
+				.Setup(i => i.GetCount())
+				.Throws(expectedException);
 
 			// Act
 			var result = _mockInner.Object.Catch((InvalidOperationException ex) =>
@@ -52,9 +48,9 @@ namespace OnionSeed.Data
 
 			// Assert
 			result.Should().NotBeNull();
-			result.Should().BeOfType<AsyncCommandServiceExceptionHandlerDecorator<FakeEntity<int>, int, InvalidOperationException>>();
+			result.Should().BeOfType<QueryServiceExceptionHandlerDecorator<FakeEntity<int>, int, InvalidOperationException>>();
 
-			var testResult = await result.TryRemoveAsync(id).ConfigureAwait(false);
+			var testResult = result.GetCount();
 			testResult.Should().Be(default);
 			actualException.Should().BeSameAs(expectedException);
